requirements.txt
django==3.2.7
djangorestframework==3.12.4
psycopg2==2.9.1                           
gunicorn==20.1.0                         
django-heroku==0.3.1
django-rest-auth==0.9.5               
django-allauth
django-rest-auth[with_social] 
